refactor: extract env.ts from orchestrator god module [EE10.02]
Move parseDotEnv, ensureEnvReady, ensureLocalEnvFile, and
loadDotEnvIntoProcess out of orchestrator.ts into a dedicated env.ts.
orchestrator.ts re-exports parseDotEnv from env.ts to preserve the public
API. Migrate parseDotEnv tests to test/env.test.ts.

ensureEnvReady accepts findPrimaryWorktreePath as an injected dependency
to avoid a circular import back to orchestrator.ts.

Zero behavior change. bun test green.
